+++ threedi_plugin.py
@@ -203,6 +203,12 @@ class ThreeDiPlugin(QObject, ProjectStateMixin):
         self.model.grid_removed.connect(self.loader.unload_grid)
         self.model.result_removed.connect(self.loader.unload_result)
 
+        # When a grid or result is loaded (or was invalid), the grid/result dialog can be enabled again
+        self.model.grid_added.connect(self.dockwidget.dialog.enable)
+        self.model.result_added.connect(self.dockwidget.dialog.enable)
+        self.validator.grid_invalid.connect(self.dockwidget.dialog.enable)
+        self.validator.result_invalid.connect(self.dockwidget.dialog.enable)
+
         self.init_state_sync()
 
         # Disable warning that scratch layer data will be lost