+++ CHANGES.rst
@@ -8,6 +8,7 @@
 - Sideview: added dots at vertical line intersections with bottom level, cross-section top, exchange level and water level) (#838)
 - Sideview: dots and vertical lines indicating nodes can now be toggled (#838)
 - Graph: removed delete button (#839)
+- Result Manager: UI is temporarily disabled when grid/result is being loaded (#860)
 
 
 3.1.5 (2023-06-21)